+++ src/main/java/nl/sander/jsontoy2/java/constantpool/MethodTypeEntry.java
@@ -1,9 +1,9 @@
 package nl.sander.jsontoy2.java.constantpool;
 
 public class MethodTypeEntry extends ConstantPoolEntry {
-    private final short descriptorIndex;
+    private final int descriptorIndex;
 
-    public MethodTypeEntry(short descriptorIndex) {
+    public MethodTypeEntry(int descriptorIndex) {
         this.descriptorIndex = descriptorIndex;
     }
